moved everything in src/test/scala so that directory matches package
