+++ src/test/scala/dsptools/tester/MemMasterSpec.scala
